requirements.txt
Package                       Version
----------------------------- ----------
alabaster                     0.7.12
asttokens                     2.1.0
Babel                         2.11.0
backcall                      0.2.0
black                         22.3.0
certifi                       2022.9.24
charset-normalizer            2.1.1
click                         8.1.3
colorama                      0.4.6
contourpy                     1.0.6
cycler                        0.11.0
decorator                     5.1.1
docutils                      0.17.1
executing                     1.2.0
flake8                        5.0.2
fonttools                     4.38.0
idna                          3.4
imagesize                     1.4.1
importlib-metadata            5.0.0
ipython                       8.6.0
jedi                          0.18.1
Jinja2                        3.1.2
joblib                        1.2.0
kiwisolver                    1.4.4
MarkupSafe                    2.1.1
matplotlib                    3.6.2
matplotlib-inline             0.1.6
mccabe                        0.7.0
mypy-extensions               0.4.3
nltk                          3.7
numpy                         1.23.5
packaging                     21.3
pandas                        1.3.3
parso                         0.8.3
pathlib2                      2.3.5
pathspec                      0.10.2
pickleshare                   0.7.5
Pillow                        9.3.0
pip                           22.3.1
platformdirs                  2.5.4
prompt-toolkit                3.0.32
pure-eval                     0.2.2
pycodestyle                   2.9.1
pydocstyle                    6.0.0
pyflakes                      2.5.0
Pygments                      2.13.0
pyparsing                     3.0.9
PyPDF2                        2.11.1
python-dateutil               2.8.2
pytz                          2022.6
PyYAML                        5.4.1
regex                         2022.10.31
requests                      2.28.1
setuptools                    65.5.1
six                           1.16.0
snowballstemmer               2.2.0
Sphinx                        4.4.0
sphinx-rtd-theme              1.0.0
sphinxcontrib-applehelp       1.0.2
sphinxcontrib-devhelp         1.0.2
sphinxcontrib-htmlhelp        2.0.0
sphinxcontrib-jsmath          1.0.1
sphinxcontrib-qthelp          1.0.3
sphinxcontrib-serializinghtml 1.1.5
stack-data                    0.6.1
tomli                         2.0.1
tqdm                          4.64.1
traitlets                     5.5.0
typing_extensions             4.4.0
urllib3                       1.26.12
wcwidth                       0.2.5
wheel                         0.37.1
wordcloud                     1.8.2.2
zipp                          3.10.0
